I realized in the previous version fifo and time_fill at the first Sell.FIFO the whole portfolio gets overwritten with the fifo() call, deleting the following transactions entirely. Added a line to save the future transactions and then append them back after the fifo(positions, sales, date) call in time_fill
Added functionality:
Calculated percentage return on a cash effective basis. Totaled cash over all time and used that as the value of the portfolio on day 0. Portfolio value then fluctuates based on the sum of unrealized and realized gains.

Whenever dividends show up in the TXN csv, add them to daily position value.